cod-whatsapp-agent/requirements.txt
# Python dependencies for COD WhatsApp Agent

flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
mysql-connector-python>=8.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
